requirements.txt
flask==1.1.1
pycryptodome==3.9.1